test: add publishing form coverage
